backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
google-generativeai>=0.7.0
duckduckgo-search==4.4.1
python-dotenv==1.0.0
pydantic==2.5.3
